close overlay when anchor scrolls out
